fixed 64-bit definitions under Linux

+++ Source/LibRawLite/internal/defines.h
@@ -56,8 +56,9 @@ it under the terms of the one of two licenses as you choose:
 #include <unistd.h>
 #include <utime.h>
 #include <netinet/in.h>
-typedef long long INT64;
-typedef unsigned long long UINT64;
+#include <stdint.h>
+typedef int64_t INT64;
+typedef uint64_t UINT64;
 #endif
 
 #ifdef NODEPS
